Stop force unwrapping rawValue
Minor update to how we are force unwrapping `rawValue`. I couldn't reproduce this consistently, it happened a few times. Specifically it happened when I was serially scanning a bunch of types that are NOT supported and then scanned one that was. I was just using barcode-generator.org as my test site. There are a few types of barcodes you'll see there, most of them are supported but there are some like Planet/Maxicode etc that are not. (#184)
Either way not force unwrapping seems like a good idea here.
There was nothing in MLKit documentation either about the feature's `rawValue` being nil.